Add xip command to open app on xip.io
xip.io is a service that forwards requests to the local IP address
specified in the domain. So requests to
http://your-pow-app.your-local-ip.xip.io will open your POW app,
assuming the computer accessing the URL has access to your local IP.
This allows one to, for instance, test a POW app from a mobile device
on the same local network.

The new command functions just like `powify browse` except that it
adds the local ip address and the xip.io extension to the URL.